+++ CHANGELOG.md
@@ -1,5 +1,3 @@
-
-
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -7,7 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [4.0.0] - 2020-08-05
+### Added:
+- Parsing functionality with `-pa` and `-n` switches
+- Groups to specify what groups to collect and/or parse with switch `-g`
+- Embadded default configuration YAML in release
 
+### Changed:
+- Changes to default configuration 
+
+### Fixed:
+- ZipFile metadata and header filename mismatch issue
 
 ## [3.2.0] - 2020-01-25